pipeline/src/extract_nlrb_run/requirements.txt
async-generator==1.10
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
chromedriver-binary==113.0.5672.24.0
chromedriver-binary-auto==0.2.5
click==8.1.3
cloudevents==1.9.0
deprecation==2.1.0
exceptiongroup==1.1.1
Flask==2.2.3
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.1
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.2
rsa==4.9
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.5.0
urllib3==1.26.15
watchdog==3.0.0
webdriver-manager==3.8.6
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
